ajuste a tipos de datos  y referencias añadidas a csproj

+++ ProyectoCore/src/Repositories/RepoClase/RepoPosicion.cs
@@ -27,7 +27,7 @@ namespace Repositories
             throw new NotImplementedException("Falta conectar el repositorio real de posiciones.");
         }
 
-        public bool Eliminar(ushort id)
+        public bool Eliminar(byte id)
         {
             throw new NotImplementedException("Falta conectar el repositorio real de posiciones.");
             return false;
